separated event processor bindings into its own module to allow for easier extension
