+++ server/src/main/java/com/netflix/conductor/server/ServerModule.java
@@ -19,6 +19,7 @@ import com.netflix.archaius.guice.ArchaiusModule;
 import com.netflix.conductor.annotations.Service;
 import com.netflix.conductor.core.config.Configuration;
 import com.netflix.conductor.core.config.CoreModule;
+import com.netflix.conductor.core.config.EventModule;
 import com.netflix.conductor.core.config.ValidationModule;
 import com.netflix.conductor.core.execution.WorkflowSweeper;
 import com.netflix.conductor.dyno.SystemPropertiesDynomiteConfiguration;
@@ -44,6 +45,7 @@ public class ServerModule extends AbstractModule {
         install(new HealthModule());
         install(new JettyModule());
         install(new GRPCModule());
+        install(new EventModule());
 
         bindInterceptor(Matchers.any(), Matchers.annotatedWith(Service.class), new ServiceInterceptor(getProvider(Validator.class)));
         bind(Configuration.class).to(SystemPropertiesDynomiteConfiguration.class);
